fix : Upgraded MainMemory property to Long to parse 32-bit (4GB) memory simulation

+++ src/config/CacheFactory.java
@@ -31,7 +31,7 @@ public class CacheFactory {
         MainMemory mainMemory = new MainMemory(); // assuming MainMemory has a default constructor (or we can inject latency/size later)
         CacheHierarchy hierarchy = new CacheHierarchy(mainMemory);
 
-        int mainMemorySize = loader.getMainMemorySize();
+        long mainMemorySize = loader.getMainMemorySize();
         int addressBits = (int) (Math.log(mainMemorySize) / Math.log(2));
 
         List<CacheConfig> configs = loader.getCacheConfigs();
